+++ Software/Cpu16Emulator/Cpu16EmulatorCpus/Tiny16v4.cs
@@ -2,7 +2,7 @@ namespace Cpu16EmulatorCpus;
 
 public class Tiny16v4(string[] code, int speed): Cpu(code, speed, 4)
 {
-    protected override ushort? IsCall(uint instruction)
+    protected override uint? IsCall(uint instruction)
     {
         return null;
     }
